requirements.txt
requests
pyyaml
flask
parse